requirements.txt
pytest==3.10.0
selenium==3.14.0
requests==2.19.1
allure-pytest==2.5.3
allure-python-commons==2.5.3
pytest-allure-adaptor==1.7.10
webdriver-manager==1.7